==> 02_NSMC/02_NSMC_RandomForest.ipynb <==
# 네이버 영화리뷰 감정분석 with Random Forest
TF-IDF 기법 참고(링크) : https://wiserloner.tistory.com/944

# 준비
라이브러리, 파라미터 세팅

import pandas as pd
import numpy as np
import os
import re

DATA_IN_PATH = './data_in/'
DATA_OUT_PATH = './data_out/'

TEST_SIZE = 0.2
RANDOM_SEED = 42

# 데이터 로드

# 학습 데이터 로딩(데이터프레임)
train_df = pd.read_csv(DATA_IN_PATH + 'ratings_train.txt', sep='\t')
train_df.head()

# 테스트(캐글) 데이터 로드
test_df = pd.read_csv(DATA_IN_PATH + 'ko_data.csv', encoding = 'cp949')
test_df.columns = ['id','document']  # 전처리 일괄 수행을 위해 컬럼명 변경(학습 데이터셋과 동일하게)
test_df.head()

# 데이터 전처리

def data_cleansing(train_data, test_data, con=0):  # 0: Null치환(공백), 1: 한글이외제거 & Null(공백)제거, 2: 숫자제거 & Null치환(공백)
    global train_df
    global test_df
    
    train_data = pd.read_csv(DATA_IN_PATH + 'ratings_train.txt', sep='\t')
    test_data = pd.read_csv(DATA_IN_PATH + 'ko_data.csv', encoding = 'cp949')
    test_data.columns = ['id','document']  # 전처리 일괄 수행을 위해 컬럼명 변경(학습 데이터셋과 동일하게)
    train_df = train_data.copy()
    test_df = test_data.copy()
    
    if con == 1:  # 한글 이외 제거 & Null(공백) 제거, row 제거 有
        train_df.drop_duplicates(subset = ['document'], inplace=True) # document 열에서 중복 내용 제거
        train_df['document'] = train_df['document'].str.replace("[^ㄱ-ㅎㅏ-ㅣ가-힣 ]","") # 정규 표현식 수행(한글 이외 제거)
        train_df['document'].replace('', np.nan, inplace=True) # 공백은 Null 값으로 변경
        train_df = train_df.dropna(how='any') # Null 값 제거(리뷰 Null은 긍/부정 의미 없음)
        
        test_df.drop_duplicates(subset = ['document'], inplace=True) # document 열에서 중복 내용 제거
        test_df['document'] = test_df['document'].str.replace("[^ㄱ-ㅎㅏ-ㅣ가-힣 ]","") # 정규 표현식 수행(한글 이외 제거)
        test_df['document'].replace('', np.nan, inplace=True) # 공백은 Null 값으로 변경
        test_df = test_df.dropna(how='any') # Null 값 제거(리뷰 Null은 긍/부정 의미 없음)
        
    elif con == 2:  # 숫자 제거 & Null 치환(공백), row 제거 없음
        # 리뷰 텍스트를 가지는 'document' 컬럼에 Null이 일부 존재하므로 이 값은 공백으로 변환
        # 정규 표현식을 이용해 숫자를 공백으로 변경(정규 표현식으로 \d는 숫자를 의미)
        train_df = train_df.fillna(' ')
        train_df['document'] = train_df['document'].apply( lambda x : re.sub(r"\d+", " ", x) )
        
        test_df = test_df.fillna(' ')
        test_df['document'] = test_df['document'].apply( lambda x : re.sub(r"\d+", " ", x) )
        
    else:
        train_df = train_df.fillna(' ')
        test_df = test_df.fillna(' ')
    
    print('전처리 후 학습 데이터 개수 :',len(train_df))
    print('전처리 후 테스트 데이터 개수 :',len(test_df))


# 데이터 전처리 선택
data_cleansing(train_df, test_df, con=0)  # 0: Null치환(공백), 1: 한글이외제거 & Null(공백)제거, 2: 숫자제거 & Null치환(공백)

train_df.head()

test_df.head()

# 학습 데이터 레이블 넘파이 배열 처리
y_train = np.array(train_df['label'])

# TF-IDF를 활용한 벡터화
TF-IDF를 통해 단어를 벡터화할 것이고, 먼저 각 문장을 한글 형태소 분석을 통해 형태소 단어로 토큰화한다. 한글 형태소 엔진은 SNS 분석에 적합한 Twitter 클래스를 이용한다. Twitter 객체의 morphs() 메소드를 이용하면 입력 인자로 들어온 문장을 형태소 단어 형태로 토큰화해 list 객체로 반환한다. 문장을 형태소 단어 형태로 반환하는 별도의 tokenizer 함수를 tw_tokenizer()라는 이름으로 생성한다. 이 함수는 사이킷런의 TfidfVectorizer 클래스의 tokenizer로 사용된다.

from konlpy.tag import Twitter

twitter = Twitter()
def tw_tokenizer(text):
    # 입력 인자로 들어온  텍스트를 형태소 단어로 토큰화해 리스트 형태로 반환
    tokens_ko = twitter.morphs(text)
    return tokens_ko

import time
from sklearn.feature_extraction.text import TfidfVectorizer

start = time.time()  # 시작 시간 저장

# TfidfVectorizer를 이용해 학습 데이터를 TF-IDF 값으로 피처 변환
# Twitter 객체의 morphs() 객체를 이용한 tokenizer를 사용. ngram_range는 (1, 2)
vectorizer = TfidfVectorizer(tokenizer=tw_tokenizer, ngram_range=(1,2), min_df=3, max_df=0.9)
vectorizer.fit(train_df['document'])
train_data_features = vectorizer.transform(train_df['document'])

# 학습 데이터를 적용한 TfidfVectorizer를 이용해 테스트 데이터를 TF-IDF 값으로 피처 변환
test_data_features = vectorizer.transform(test_df['document'])

print("time :", time.time() - start)  # 현재시각 - 시작시간 = 실행 시간
print()
print(train_data_features.shape)
print(test_data_features.shape)
print(type(train_data_features))

# (Baseline) 학습과 검증 데이터 분리 후 모델 학습

from sklearn.model_selection import train_test_split

train_input, eval_input, train_label, eval_label = train_test_split(train_data_features, y_train, test_size=TEST_SIZE, random_state=RANDOM_SEED)

from sklearn.ensemble import RandomForestClassifier

start = time.time()  # 시작 시간 저장

# 랜덤 포레스트 사용해서 감성 분석 분류 수행
clf = RandomForestClassifier(random_state = RANDOM_SEED)  # clf = RandomForestClassifier(n_estimators = 100)
clf.fit(train_input, train_label)

print("time :", time.time() - start)  # 현재시각 - 시작시간 = 실행 

print("Accuracy: %f" % clf.score(eval_input, eval_label))  # 검증 함수로 정확도 측정

# 모델(GridSearch) 구현 및 학습 

'''
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import GridSearchCV

start = time.time()  # 시작 시간 저장

# 랜덤 포레스트 사용해서 감성 분석 분류 수행
rf_clf = RandomForestClassifier(random_state = RANDOM_SEED)

# 파라미터 최적화를 위해 GridSearchCV를 이용
# params = { 'n_estimators' : [10, 100],
#            'max_depth' : [6, 8, 10, 12],
#            'min_samples_leaf' : [8, 12, 18],
#            'min_samples_split' : [8, 16, 20] }
params = {}

grid_cv = GridSearchCV(rf_clf, param_grid=params, cv=5, scoring='accuracy', verbose=1, n_jobs=32)
grid_cv.fit(train_data_features, y_train)
print(grid_cv.best_params_, round(grid_cv.best_score_, 4))

print("time :", time.time() - start)  # 현재시각 - 시작시간 = 실행 
'''

# Original 테스트 데이터셋으로 성능 평가

# 검증 데이터 로드
dev_df = pd.read_csv(DATA_IN_PATH + 'ratings_test.txt', sep='\t')
dev_df.head()

# 리뷰 텍스트를 가지는 'document' 컬럼에 Null이 일부 존재하므로 이 값은 공백으로 변환
dev_df = dev_df.fillna(' ')
# # 정규 표현식을 이용해 숫자를 공백으로 변경(정규 표현식으로 \d는 숫자를 의미)
# dev_df['document'] = dev_df['document'].apply( lambda x : re.sub(r"\d+", " ", x) )

dev_df.head()

# 학습 데이터를 적용한 TfidfVectorizer를 이용해 검증 데이터를 TF-IDF 값으로 피처 변환
dev_data_features = vectorizer.transform(dev_df['document'])

from sklearn.metrics import accuracy_score, precision_score, recall_score, confusion_matrix
from sklearn.metrics import f1_score, roc_auc_score

# # classifier는 GridSearchCV에서 최적 파라미터로 학습된 classifier를 그대로 이용
# best_estimator = grid_cv.best_estimator_
# preds = best_estimator.predict(test_data_features)

preds = clf.predict(dev_data_features)

confusion = confusion_matrix(dev_df['label'], preds)

print('오차 행렬')
print(confusion,'\n')

print('Random Forest 정확도  :', accuracy_score(dev_df['label'], preds))
print('Random Forest 정밀도  :', precision_score(dev_df['label'], preds))
print('Random Forest 재현율  :', recall_score(dev_df['label'], preds))
print('Random Forest F1스코어:', f1_score(dev_df['label'], preds))
print('Random Forest AUC     :', roc_auc_score(dev_df['label'], preds))

# (참고) 제출 파일 생성

# 테스트 데이터의 리뷰 부분을 리스트 처리
test_reviews = list(test_df['id'])

test_preds = clf.predict(test_data_features)

# 판다스 데이터프레임 통해 데이터 구성하여 output에 투입
output = pd.DataFrame( data={"Id": test_reviews, "Predicted": test_preds} )
output.head()

# 해당 경로가 없으면 생성
if not os.path.exists(DATA_OUT_PATH):
    os.makedirs(DATA_OUT_PATH)

# csv파일로 만든다
output.to_csv( DATA_OUT_PATH + "NSMC_RandomForest.csv", index = False)

### 캐글 제출 결과
**[2020.12.20]**<br>
0.81745

# (참고) 리뷰 예측하기

# 입력을 리스트 형태로 넣어야 함
# 참고 : https://m.blog.naver.com/PostView.nhn?blogId=samsjang&logNo=220985170721&proxyReferer=https:%2F%2Fwww.google.com%2F
def sentiment_predict(new_sentence):
    label = {0: '부정', 1: '긍정'}
    new_sentence = vectorizer.transform(new_sentence)  # TF-IDF 값으로 피처 변환
    print('예측: %s\n확률: %.3f%%' %(label[clf.predict(new_sentence)[0]], np.max(clf.predict_proba(new_sentence))*100))

sentiment_predict(['이 영화 개꿀잼 ㅋㅋㅋ'])

sentiment_predict(['이 영화 핵노잼 ㅠㅠ'])

sentiment_predict(['견디기 힘드네요'])

---
