==> 02_NSMC/09_NSMC_KoBERT_Pytorch.ipynb <==
# 네이버 영화리뷰 감정분석 with SKT KoBERT
참고 소스 출처(링크) : https://github.com/SKTBrain/KoBERT

https://github.com/SKTBrain/KoBERT#using-with-pytorch

https://colab.research.google.com/github/SKTBrain/KoBERT/blob/master/scripts/NSMC/naver_review_classifications_pytorch_kobert.ipynb

# 준비
라이브러리, 파라미터 세팅

# !pip install mxnet
# !pip install gluonnlp pandas tqdm
# !pip install sentencepiece
# !pip install transformers
# !pip install torch

# !pip install git+https://git@github.com/SKTBrain/KoBERT.git@master

import torch
from torch import nn
import torch.nn.functional as F
import torch.optim as optim
from torch.utils.data import Dataset, DataLoader
import gluonnlp as nlp
import numpy as np
import pandas as pd
from tqdm import tqdm, tqdm_notebook
import csv
import os

from kobert.utils import get_tokenizer
from kobert.pytorch_kobert import get_pytorch_kobert_model

from transformers import AdamW
from transformers.optimization import get_cosine_schedule_with_warmup

### Parameter Setting

DATA_IN_PATH = './data_in/'
DATA_OUT_PATH = './data_out/'

TEST_SIZE = 0.2
RANDOM_SEED = 42
MAXLEN = 128  # 64
BATCHSIZE = 32  # 64
EPOCHS = 5  # 5

### GPU Setting

# 디바이스 설정 확인
if torch.cuda.is_available():    
    device = torch.device("cuda")
    print('There are %d GPU(s) available.' % torch.cuda.device_count())
    print('We will use the GPU:', torch.cuda.get_device_name(0))
else:
    device = torch.device("cpu")
    print('No GPU available, using the CPU instead.')

# GPU 할당 변경하기
GPU_NUM = 2 # 원하는 GPU 번호 입력
device = torch.device(f'cuda:{GPU_NUM}' if torch.cuda.is_available() else 'cpu')
torch.cuda.set_device(device) # change allocation of current GPU
print ('Current cuda device ', torch.cuda.current_device()) # check

### KoBERT Setting

bertmodel, vocab = get_pytorch_kobert_model()

tokenizer = get_tokenizer()
tok = nlp.data.BERTSPTokenizer(tokenizer, vocab, lower=False)

# 데이터 로드

# 학습 데이터 로드
train = pd.read_csv(DATA_IN_PATH + 'ratings_train.txt', sep='\t')
train.head()

# 검증(Original Test Dataset) 데이터 로드
dev = pd.read_csv(DATA_IN_PATH + 'ratings_test.txt', sep='\t')
dev.head()

# 테스트(캐글) 데이터 로드
test = pd.read_csv(DATA_IN_PATH + 'ko_data.csv', encoding = 'cp949')
test.columns = ['id','document']
test.head()

# 학습/검증 데이터를 txt 파일로 저장
train.to_csv(DATA_IN_PATH+'train.txt', sep = '\t', index = True)
dev.to_csv(DATA_IN_PATH+'dev.txt', sep = '\t', index = True)

# 테스트(캐글) 데이터를 txt 파일로 저장(label 없음)
csv_file = './data_in/ko_data.csv'
txt_file = './data_in/ko_data.txt'
with open(txt_file, "w") as my_output_file:
    with open(csv_file, "r", encoding='cp949') as my_input_file: # , encoding='cp949'
        [ my_output_file.write("\t".join(row)+'\n') for row in csv.reader(my_input_file)]
    my_output_file.close()

dataset_train= nlp.data.TSVDataset(DATA_IN_PATH+"train.txt", field_indices=[2,3], num_discard_samples=1)
dataset_dev = nlp.data.TSVDataset(DATA_IN_PATH+"dev.txt", field_indices=[2,3], num_discard_samples=1)

dataset_test = nlp.data.TSVDataset(DATA_IN_PATH+"ko_data.txt", field_indices=[1], num_discard_samples=1)

dataset_test

# 모델 구현

class BERTDataset(Dataset):
    def __init__(self, dataset, sent_idx, label_idx, bert_tokenizer, max_len,
                 pad, pair):
        transform = nlp.data.BERTSentenceTransform(
            bert_tokenizer, max_seq_length=max_len, pad=pad, pair=pair)

        self.sentences = [transform([i[sent_idx]]) for i in dataset]
        self.labels = [np.int32(i[label_idx]) for i in dataset]

    def __getitem__(self, i):
        return (self.sentences[i] + (self.labels[i], ))

    def __len__(self):
        return (len(self.labels))

class BERTDataset_Test(Dataset):
    def __init__(self, dataset, sent_idx, bert_tokenizer, max_len,
                 pad, pair):
        transform = nlp.data.BERTSentenceTransform(
            bert_tokenizer, max_seq_length=max_len, pad=pad, pair=pair)

        self.sentences = [transform([i[sent_idx]]) for i in dataset]
        #self.labels = [np.int32(i[label_idx]) for i in dataset]

    def __getitem__(self, i):
        return (self.sentences[i])

    def __len__(self):
        return (len(self.sentences))

## Setting parameters
max_len = MAXLEN  # 64
batch_size = BATCHSIZE  # 64
warmup_ratio = 0.1
num_epochs = EPOCHS  # 5
max_grad_norm = 1
log_interval = 200
learning_rate =  5e-5

data_train = BERTDataset(dataset_train, 0, 1, tok, max_len, True, False)
data_dev = BERTDataset(dataset_dev, 0, 1, tok, max_len, True, False)
data_test = BERTDataset_Test(dataset_test, 0, tok, max_len, True, False)

train_dataloader = torch.utils.data.DataLoader(data_train, batch_size=batch_size, num_workers=5)
dev_dataloader = torch.utils.data.DataLoader(data_dev, batch_size=batch_size, num_workers=5)
test_dataloader = torch.utils.data.DataLoader(data_test, batch_size=batch_size, num_workers=5)

class BERTClassifier(nn.Module):
    def __init__(self,
                 bert,
                 hidden_size = 768,
                 num_classes=2,
                 dr_rate=None,
                 params=None):
        super(BERTClassifier, self).__init__()
        self.bert = bert
        self.dr_rate = dr_rate
                 
        self.classifier = nn.Linear(hidden_size , num_classes)
        if dr_rate:
            self.dropout = nn.Dropout(p=dr_rate)
    
    def gen_attention_mask(self, token_ids, valid_length):
        attention_mask = torch.zeros_like(token_ids)
        for i, v in enumerate(valid_length):
            attention_mask[i][:v] = 1
        return attention_mask.float()

    def forward(self, token_ids, valid_length, segment_ids):
        attention_mask = self.gen_attention_mask(token_ids, valid_length)
        
        _, pooler = self.bert(input_ids = token_ids, token_type_ids = segment_ids.long(), attention_mask = attention_mask.float().to(token_ids.device))
        if self.dr_rate:
            out = self.dropout(pooler)
        return self.classifier(out)

model = BERTClassifier(bertmodel, dr_rate=0.5).to(device)

# Prepare optimizer and schedule (linear warmup and decay)
no_decay = ['bias', 'LayerNorm.weight']
optimizer_grouped_parameters = [
    {'params': [p for n, p in model.named_parameters() if not any(nd in n for nd in no_decay)], 'weight_decay': 0.01},
    {'params': [p for n, p in model.named_parameters() if any(nd in n for nd in no_decay)], 'weight_decay': 0.0}
]

optimizer = AdamW(optimizer_grouped_parameters, lr=learning_rate)
loss_fn = nn.CrossEntropyLoss()

t_total = len(train_dataloader) * num_epochs
warmup_step = int(t_total * warmup_ratio)

scheduler = get_cosine_schedule_with_warmup(optimizer, num_warmup_steps=warmup_step, num_training_steps=t_total)

# 모델 학습 및 평가

def calc_accuracy(X,Y):
    max_vals, max_indices = torch.max(X, 1)
    train_acc = (max_indices == Y).sum().data.cpu().numpy()/max_indices.size()[0]
    return train_acc

for e in range(num_epochs):
    train_acc = 0.0
    test_acc = 0.0
    
    print()
    print('======== Epoch {:} / {:} ========'.format(e+1, num_epochs))
    
    # 모델 학습
    model.train()
    for batch_id, (token_ids, valid_length, segment_ids, label) in enumerate(tqdm_notebook(train_dataloader)):
        optimizer.zero_grad()
        token_ids = token_ids.long().to(device)
        segment_ids = segment_ids.long().to(device)
        valid_length= valid_length
        label = label.long().to(device)
        out = model(token_ids, valid_length, segment_ids)
        loss = loss_fn(out, label)
        loss.backward()
        torch.nn.utils.clip_grad_norm_(model.parameters(), max_grad_norm)
        optimizer.step()
        scheduler.step()  # Update learning rate schedule
        train_acc += calc_accuracy(out, label)
        if batch_id % log_interval == 0:
            print("epoch {} batch id {} loss {} train acc {}".format(e+1, batch_id+1, loss.data.cpu().numpy(), train_acc / (batch_id+1)))
    print("epoch {} train acc {}".format(e+1, train_acc / (batch_id+1)))
    
#     # 모델 평가 => GPU 메모리 최적화 위해 주석 처리
#     model.eval()
#     for batch_id, (token_ids, valid_length, segment_ids, label) in enumerate(tqdm_notebook(dev_dataloader)):
#         token_ids = token_ids.long().to(device)
#         segment_ids = segment_ids.long().to(device)
#         valid_length= valid_length
#         label = label.long().to(device)
#         out = model(token_ids, valid_length, segment_ids)
#         test_acc += calc_accuracy(out, label)
#     print("epoch {} test acc {}".format(e+1, test_acc / (batch_id+1)))
    print()

### F1-Score 확인

dev_preds = []
y_dev = []

for batch_id, (token_ids, valid_length, segment_ids, label) in enumerate(tqdm_notebook(dev_dataloader)):
    token_ids = token_ids.long().to(device)
    segment_ids = segment_ids.long().to(device)
    valid_length = valid_length
    label = label.to('cpu').numpy()  # label을 torch tensor에서 numpy 타입으로 변경
    out = model(token_ids, valid_length, segment_ids)
    
    for i in label:
        y_dev.append(i)
    
    for i in range(len(out)):
        dev_preds.append(np.argmax(out[i].detach().cpu().numpy()))

dev_preds[:5], y_dev[:5]

from sklearn.metrics import f1_score
f1 = f1_score(y_dev, dev_preds, average='weighted')
print('F1 Score : {:.5f}'.format(f1))

# (참고) 제출 파일 생성

test_preds = []

for batch_id, (token_ids, valid_length, segment_ids) in enumerate(tqdm_notebook(test_dataloader)):
    token_ids = token_ids.long().to(device)
    segment_ids = segment_ids.long().to(device)
    valid_length = valid_length
    out = model(token_ids, valid_length, segment_ids)
    
    for i in range(len(out)):
        test_preds.append(np.argmax(out[i].detach().cpu().numpy()))

len(test_preds)

test_preds[:5]

# 테스트 데이터의 리뷰 부분을 리스트 처리
test_id = list(test['id'])

# 데이터프레임 통해 데이터 구성하여 output에 투입
output = pd.DataFrame( data={"Id": test_id, "Predicted": test_preds} )
output.head()

# 해당 경로가 없으면 생성
if not os.path.exists(DATA_OUT_PATH):
    os.makedirs(DATA_OUT_PATH)

SAVE_NM = "NSMC_KoBERT_MAXLN"+str(max_len)+"_BATSZ"+str(batch_size)+"_EPOCH"+str(num_epochs)+".csv"

# csv 파일 생성
output.to_csv(DATA_OUT_PATH + SAVE_NM, index = False)  # 앙상블 조합 및 캐글 제출 용도

### 캐글 제출 결과
**[2020.12.14]**<br>
max_len = 64, batch_size = 64, epoch 1 => 0.87162<br>
max_len = 64, batch_size = 64, epoch 3 => 0.89719<br>
**max_len = 64, batch_size = 64, epoch 5 => 0.90076** => 0.90452<br>

max_len = 64, batch_size = 32, epoch 1 => 0.87681<br>
max_len = 64, batch_size = 32, epoch 3 => 0.88896<br>
**max_len = 64, batch_size = 32, epoch 5 => 0.90005** => 0.90059<br>

max_len = 128, batch_size = 32, epoch 1 => 0.87806<br>
max_len = 128, batch_size = 32, epoch 3 => 0.89504<br>
**max_len = 128, batch_size = 32, epoch 5 => 0.90112** => 0.90541<br>

(실패) max_len = 128, batch_size = 64, epoch 1 => Out of Memory (GPU)<br>
(실패) max_len = 128, batch_size = 64, epoch 3 => Out of Memory (GPU)<br>
(실패) max_len = 128, batch_size = 64, epoch 5 => Out of Memory (GPU)<br>

**[2020.12.15]**<br>
**max_len = 64, batch_size = 32, epoch 5 => 0.90059<br>
max_len = 64, batch_size = 64, epoch 5 => 0.90452<br>
max_len = 128, batch_size = 32, epoch 5 => 0.90541**<br>
