==> 02_NSMC/10_NSMC_KoELECTRA_Pytorch_SentiWordDict_KHS.ipynb <==
# 네이버 영화리뷰 감정분석 with Hugging Face KoELECTRA and Data Augmentation(KNU 감성사전 + 한국어 혐오발언)
참고 소스 출처(링크) : https://github.com/monologg/KoELECTRA<br>
https://blog.naver.com/horajjan/221739630055<br>
https://bab2min.tistory.com/657

# 준비
라이브러리 , 파라미터 세팅

import tensorflow as tf
import torch

from transformers import ElectraTokenizer
from transformers import ElectraForSequenceClassification, AdamW
from transformers import get_linear_schedule_with_warmup
from torch.utils.data import TensorDataset, DataLoader, RandomSampler, SequentialSampler
from keras.preprocessing.sequence import pad_sequences
from sklearn.model_selection import train_test_split

import pandas as pd
import numpy as np
import random
import time
import datetime
import os

DATA_IN_PATH = './data_in/'
DATA_OUT_PATH = './data_out/'

TEST_SIZE = 0.2
RANDOM_SEED = 42
EPOCHS = 4  # 3

# 디바이스 설정 확인
if torch.cuda.is_available():    
    device = torch.device("cuda")
    print('There are %d GPU(s) available.' % torch.cuda.device_count())
    print('We will use the GPU:', torch.cuda.get_device_name(0))
else:
    device = torch.device("cpu")
    print('No GPU available, using the CPU instead.')

# GPU 할당 변경하기
GPU_NUM = 3 # 원하는 GPU 번호 입력
device = torch.device(f'cuda:{GPU_NUM}' if torch.cuda.is_available() else 'cpu')
torch.cuda.set_device(device) # change allocation of current GPU
print ('Current cuda device ', torch.cuda.current_device()) # check

# 데이터 로드

# 학습 데이터 로드
train = pd.read_csv(DATA_IN_PATH + 'ratings_train.txt', sep='\t')
train.head()

# 검증(Original Test Dataset) 데이터 로드
dev = pd.read_csv(DATA_IN_PATH + 'ratings_test.txt', sep='\t')
dev.head()

# 테스트(캐글) 데이터 로드
test = pd.read_csv(DATA_IN_PATH + 'ko_data.csv', encoding = 'cp949')
test.columns = ['id','document']  # 전처리 일괄 수행을 위해 컬럼명 변경(학습 데이터셋과 동일하게)
test.head()

train.drop('id', axis=1, inplace=True)
#dev.drop('id', axis=1, inplace=True)

### 추가 데이터 로딩 (KNU 한국어 감성사전)
https://github.com/park1200656/KnuSentiLex

본 한국어 감성사전은 다음과 같은 소스로부터 통합되어 개발되었음<br>
(1) 국립국어원 표준국어대사전의 뜻풀이(glosses) 분석을 통한 긍부정 추출(이 방법을 통해 대부분의 긍부정어 추출)<br>
(2) 김은영(2004)의 긍부정어 목록<br>
(3) SentiWordNet 및 SenticNet-5.0에서 주로 사용되는 긍부정어 번역<br>
(4) 최근 온라인에서 많이 사용되는 축약어 및 긍부정 이모티콘 목록<br>

총 14,843개의 1-gram, 2-gram, 관용구, 문형, 축약어, 이모티콘 등에 대한 긍정, 중립, 부정 판별 및 정도(degree)값 계산

# 추가 데이터(naver_shopping.txt) 로딩 (Naver Shopping Review)
train_plus_df1 = pd.read_csv(DATA_IN_PATH + 'SentiWordDict/' + 'SentiWord_Dict.txt', sep='\t', header = None)  # 데이터프레임 로드
train_plus_df1.columns = ['document','label']  # 컬럼명 부여
train_plus_df1.head()

# 전체 데이터 개수
len(train_plus_df1)

# 레이블 현황 확인
train_plus_df1['label'].value_counts()

# '0.0'(중립 감정) 레이블 제거
idx_neutral = train_plus_df1[train_plus_df1['label'] == 0.0].index
train_plus_df1 = train_plus_df1.drop(idx_neutral)
len(train_plus_df1)

# 레이블 값 변경
train_plus_df1.label[train_plus_df1.label == -1.0] = 0  # -1.0 => 부정 0
train_plus_df1.label[train_plus_df1.label == -2.0] = 0  # -2.0 => 부정 0
train_plus_df1.label[train_plus_df1.label == 1.0] = 1  # 1.0 => 긍정 1
train_plus_df1.label[train_plus_df1.label == 2.0] = 1  # 2.0 => 긍정 1
train_plus_df1['label'].value_counts()

# null 값 여부 확인
print(train_plus_df1['label'].isnull().sum())

# null 값 row 제거
train_plus_df1.dropna(axis=0, inplace = True)

train_plus_df1.head()

print(train_plus_df1.dtypes)

# 레이블 데이터 타입 변환(float64->int64)
train_plus_df1 = train_plus_df1.astype({'label': 'int64'})

print(train_plus_df1.dtypes)

plus_result_tmp = pd.concat([train,train_plus_df1], axis=0, ignore_index=True)
train = plus_result_tmp.copy()
train

### 추가 데이터 로딩 (Korean Hate Speech)
https://github.com/monologg/korean-hate-speech-koelectra

한국어 혐오발언 분류 데이터셋

# 추가 데이터(train.txt) 로딩1 (Korean Hate Speech)
train_plus_df2 = pd.read_csv(DATA_IN_PATH + 'korean_hate_speech/' + 'train.txt', sep='\t')

# 추가 데이터(validate.txt) 로딩2 (Korean Hate Speech)
validation_plus_df = pd.read_csv(DATA_IN_PATH + 'korean_hate_speech/' + 'validate.txt', sep='\t')

train_plus_df2.head()

# 데이터 타입 확인
print(train_plus_df2.dtypes)
print(validation_plus_df.dtypes)

train_plus_df2.drop('title', axis=1, inplace=True)
train_plus_df2.drop('bias', axis=1, inplace=True)
validation_plus_df.drop('title', axis=1, inplace=True)
validation_plus_df.drop('bias', axis=1, inplace=True)

train_plus_df2.columns = ['document','label']
validation_plus_df.columns = ['document','label']

train_plus_df2.label[train_plus_df2.label == 'hate'] = 0
validation_plus_df.label[validation_plus_df.label == 'hate'] = 0

df1 = train_plus_df2[train_plus_df2['label']==0]
df2 = validation_plus_df[validation_plus_df['label']==0]

plus_result = pd.concat([df1,df2], axis=0, ignore_index=True)
plus_result

plus_result_tmp = pd.concat([train,plus_result], axis=0, ignore_index=True)
train = plus_result_tmp.copy()
train

# 테이터 shape 확인
print(train.shape)
print(dev.shape)
print(test.shape)

# null 값 여부 확인
print(train['label'].isnull().sum())

# label 컬럼 데이터 타입 변경 (object -> int64)
train = train.astype({'label': 'int64'})

# 데이터 타입 확인
print(train.dtypes)
print(dev.dtypes)
print(test.dtypes)

# 데이터 전처리

MAX_LEN = 128

def getInputs(dataset):
    data = dataset.copy(deep=True)
    
    if 'document' in data.columns:
        sentences = data['document']
    else:
        sentences = data['Sentence']
    sentences = ["[CLS] " + str(sentence) + " [SEP]" for sentence in sentences]
    
    tokenizer = ElectraTokenizer.from_pretrained("monologg/koelectra-base-v3-discriminator", do_lower_case=False)
    tokenized_texts = [tokenizer.tokenize(sent) for sent in sentences]
    
    input_ids = [tokenizer.convert_tokens_to_ids(x) for x in tokenized_texts]
    input_ids = pad_sequences(input_ids, maxlen=MAX_LEN, dtype="long", truncating="post", padding="post")
    
    attention_masks = []
    for seq in input_ids:
        seq_mask = [float(i>0) for i in seq]
        attention_masks.append(seq_mask)
    
    return input_ids, attention_masks

def getIndex(dataset):
    data = dataset.copy(deep = True)
    input_index = data.index.tolist()
    return torch.tensor(input_index)

# 훈련셋,검증셋 분리 전 학습 데이터셋 라벨 추출
labels = train['label'].values
# 검증(Original Test Dataset) 데이터셋 라벨 추출
dev_labels = dev['label'].values

ratings_inputs, ratings_masks = getInputs(train)
test_inputs, test_masks = getInputs(test)
dev_inputs, dev_masks = getInputs(dev)

# 훈련셋과 검증셋으로 분리
train_inputs, validation_inputs, train_labels, validation_labels = train_test_split(ratings_inputs, labels, random_state=RANDOM_SEED, test_size=TEST_SIZE)

# 어텐션 마스크를 훈련셋과 검증셋으로 분리
train_masks, validation_masks, _, _ = train_test_split(ratings_masks, ratings_inputs, random_state=RANDOM_SEED, test_size=TEST_SIZE)

# 학습셋 텐서 변환(검증 포함)
# 학습/검증/테스트 데이터를 파이토치의 텐서로 변환
train_inputs = torch.tensor(train_inputs)
train_labels = torch.tensor(train_labels)
train_masks = torch.tensor(train_masks)
validation_inputs = torch.tensor(validation_inputs)
validation_labels = torch.tensor(validation_labels)
validation_masks = torch.tensor(validation_masks)

# 테스트셋(Kaggle Dataset) 텐서 변환
test_index = getIndex(test)
test_inputs = torch.tensor(test_inputs)
test_masks = torch.tensor(test_masks)

# 검증용(Original Test Dataset) 텐서 변환
dev_inputs = torch.tensor(dev_inputs)
dev_labels = torch.tensor(dev_labels)
dev_masks = torch.tensor(dev_masks)

batch_size = 32

# 파이토치의 DataLoader로 입력, 마스크, 라벨을 묶어 데이터 설정
# 학습시 배치 사이즈 만큼 데이터를 가져옴
train_data = TensorDataset(train_inputs, train_masks, train_labels)
train_sampler = RandomSampler(train_data)
train_dataloader = DataLoader(train_data, sampler=train_sampler, batch_size=batch_size)

validation_data = TensorDataset(validation_inputs, validation_masks, validation_labels)
validation_sampler = SequentialSampler(validation_data)
validation_dataloader = DataLoader(validation_data, sampler=validation_sampler, batch_size=batch_size)

# 테스트셋(Kaggle Dataset)
test_data = TensorDataset(test_index, test_inputs, test_masks)
test_sampler = RandomSampler(test_data)
test_dataloader = DataLoader(test_data, sampler=test_sampler, batch_size=batch_size)

# 검증용(Original Test Dataset)
dev_data = TensorDataset(dev_inputs, dev_masks, dev_labels)
dev_sampler = RandomSampler(dev_data)
dev_dataloader = DataLoader(dev_data, sampler=dev_sampler, batch_size=batch_size)

# 모델 생성

# ELECTRA 모델 생성 (koelectra-base-v3-discriminator)
model = ElectraForSequenceClassification.from_pretrained("monologg/koelectra-base-v3-discriminator", num_labels = 2)
model.cuda()

# 옵티마이저 설정
optimizer = AdamW(model.parameters(),
                  lr = 3e-5, # 학습률
                  eps = 1e-8 # 0으로 나누는 것을 방지하기 위한 epsilon 값
                )

# 에폭수
epochs = EPOCHS

# 총 훈련 스텝 : 배치반복 횟수 * 에폭
total_steps = len(train_dataloader) * epochs

# 학습률을 조금씩 감소시키는 스케줄러 생성
scheduler = get_linear_schedule_with_warmup(optimizer, 
                                            num_warmup_steps = 0,
                                            num_training_steps = total_steps)

# 모델 학습

# 정확도 계산 함수
def flat_accuracy(preds, labels):
    
    pred_flat = np.argmax(preds, axis=1).flatten()
    labels_flat = labels.flatten()

    return np.sum(pred_flat == labels_flat) / len(labels_flat)

# 시간 표시 함수
def format_time(elapsed):

    # 반올림
    elapsed_rounded = int(round((elapsed)))
    
    # hh:mm:ss으로 형태 변경
    return str(datetime.timedelta(seconds=elapsed_rounded))

# 재현을 위해 랜덤시드 고정
seed_val = RANDOM_SEED
random.seed(seed_val)
np.random.seed(seed_val)
torch.manual_seed(seed_val)
torch.cuda.manual_seed_all(seed_val)

# 그래디언트 초기화
model.zero_grad()

# 에폭만큼 반복
for epoch_i in range(0, epochs):
    
    # ========================================
    #               Training
    # ========================================
    
    print("")
    print('======== Epoch {:} / {:} ========'.format(epoch_i + 1, epochs))
    print('Training...')

    # 시작 시간 설정
    t0 = time.time()

    # 로스 초기화
    total_loss = 0

    # 훈련모드로 변경
    model.train()
        
    # 데이터로더에서 배치만큼 반복하여 가져옴
    for step, batch in enumerate(train_dataloader):
        # 경과 정보 표시
        if step % 500 == 0 and not step == 0:
            elapsed = format_time(time.time() - t0)
            print('  Batch {:>5,}  of  {:>5,}.    Elapsed: {:}.'.format(step, len(train_dataloader), elapsed))

        # 배치를 GPU에 넣음
        batch = tuple(t.to(device) for t in batch)
        
        # 배치에서 데이터 추출
        b_input_ids, b_input_mask, b_labels = batch

        # Forward 수행                
        outputs = model(b_input_ids, 
                        token_type_ids=None, 
                        attention_mask=b_input_mask, 
                        labels=b_labels)
        
        # 로스 구함
        loss = outputs[0]

        # 총 로스 계산
        total_loss += loss.item()

        # Backward 수행으로 그래디언트 계산
        loss.backward()

        # 그래디언트 클리핑
        torch.nn.utils.clip_grad_norm_(model.parameters(), 1.0)

        # 그래디언트를 통해 가중치 파라미터 업데이트
        optimizer.step()

        # 스케줄러로 학습률 감소
        scheduler.step()

        # 그래디언트 초기화
        model.zero_grad()

    # 평균 로스 계산
    avg_train_loss = total_loss / len(train_dataloader)            

    print("")
    print("  Average training loss: {0:.5f}".format(avg_train_loss))
    print("  Training epcoh took: {:}".format(format_time(time.time() - t0)))
        
    # ========================================
    #               Validation
    # ========================================

    print("")
    print("Running Validation...")

    #시작 시간 설정
    t0 = time.time()

    # 평가모드로 변경
    model.eval()

    # 변수 초기화
    eval_loss, eval_accuracy = 0, 0
    nb_eval_steps, nb_eval_examples = 0, 0

    # 데이터로더에서 배치만큼 반복하여 가져옴
    for batch in validation_dataloader:
        # 배치를 GPU에 넣음
        batch = tuple(t.to(device) for t in batch)
        
        # 배치에서 데이터 추출
        b_input_ids, b_input_mask, b_labels = batch
        
        # 그래디언트 계산 안함
        with torch.no_grad():     
            # Forward 수행
            outputs = model(b_input_ids, 
                            token_type_ids=None, 
                            attention_mask=b_input_mask)
        
        # 로스 구함
        logits = outputs[0]

        # CPU로 데이터 이동
        logits = logits.detach().cpu().numpy()
        label_ids = b_labels.to('cpu').numpy()
        
        # 출력 로직과 라벨을 비교하여 정확도 계산
        tmp_eval_accuracy = flat_accuracy(logits, label_ids)
        eval_accuracy += tmp_eval_accuracy
        nb_eval_steps += 1

    print("  Accuracy: {0:.5f}".format(eval_accuracy/nb_eval_steps))
    print("  Validation took: {:}".format(format_time(time.time() - t0)))

print("")
print("Training complete!")

# 검증 데이터셋(Original Test Dataset) 평가

#시작 시간 설정
t0 = time.time()

# 평가모드로 변경
model.eval()

# 변수 초기화
eval_loss, eval_accuracy = 0, 0
nb_steps, nb_examples = 0, 0

# 데이터로더에서 배치만큼 반복하여 가져옴
for step, batch in enumerate(dev_dataloader):
    # 경과 정보 표시
    if step % 100 == 0 and not step == 0:
        elapsed = format_time(time.time() - t0)
        print('  Batch {:>5,}  of  {:>5,}.    Elapsed: {:}.'.format(step, len(dev_dataloader), elapsed))

    # 배치를 GPU에 넣음
    batch = tuple(t.to(device) for t in batch)
    
    # 배치에서 데이터 추출
    b_input_ids, b_input_mask, b_labels = batch
    
    # 그래디언트 계산 안함
    with torch.no_grad():     
        # Forward 수행
        outputs = model(b_input_ids, 
                        token_type_ids=None, 
                        attention_mask=b_input_mask)
    
    # 로스 구함
    logits = outputs[0]

    # CPU로 데이터 이동
    logits = logits.detach().cpu().numpy()
    label_ids = b_labels.to('cpu').numpy()
    
    # 출력 로짓과 라벨을 비교하여 정확도 계산
    tmp_accuracy = flat_accuracy(logits, label_ids)
    eval_accuracy += tmp_accuracy
    nb_steps += 1

print("")
print("Accuracy: {0:.5f}".format(eval_accuracy/nb_steps))
print("Test took: {:}".format(format_time(time.time() - t0)))

# (참고) 제출 파일 생성

tmp_test_dataloader = DataLoader(test_data, sampler=test_sampler, batch_size=1)
test_result = test.copy(deep = True)
test_result['Predicted'] = 'default'
classes = [0, 1]

#시작 시간 설정
t0 = time.time()

# 평가모드로 변경
model.eval()

# 변수 초기화
nb_eval_steps, nb_eval_examples = 0, 0

# 데이터로더에서 배치만큼 반복하여 가져옴
for step, batch in enumerate(tmp_test_dataloader):
    # 경과 정보 표시
    if step % 100 == 0 and not step == 0:
        elapsed = format_time(time.time() - t0)
        print('  Batch {:>5,}  of  {:>5,}.    Elapsed: {:}.'.format(step, len(test_dataloader), elapsed))

    # 배치를 GPU에 넣음
    batch = tuple(t.to(device) for t in batch)
    
    # 배치에서 데이터 추출
    b_index, b_input_ids, b_input_mask = batch
    
    # 그래디언트 계산 안함
    with torch.no_grad():     
        # Forward 수행
        outputs = model(b_input_ids, 
                        token_type_ids=None, 
                        attention_mask=b_input_mask)
    
    # 로스 구함
    logits = outputs[0]

    # CPU로 데이터 이동
    logits = logits.detach().cpu().numpy()
    idx = b_index.item()
    test_result['Predicted'][idx] = classes[np.argmax(logits)]
    
    nb_eval_steps += 1

print("")
print("Test took: {:}".format(format_time(time.time() - t0)))

test_result.head()

# 모델 산출물(데이터프레임) 수정
test_result.drop('document', axis=1, inplace=True)
test_result = test_result.rename({'id':'Id'}, axis='columns')
test_result

# 해당 경로가 없으면 생성
if not os.path.exists(DATA_OUT_PATH):
    os.makedirs(DATA_OUT_PATH)

# csv파일 생성
test_result.to_csv(DATA_OUT_PATH + "NSMC_KoELECTRA_SentiWordDict_KHS.csv", index = False)  # 앙상블 조합 및 캐글 제출

### 캐글 제출 결과
**[2020.12.10]**<br>
epoch 3 => 0.89379<br>
epoch 4 => 0.89880 (best)

---
