==> 02_NSMC/12_NSMC_Ensemble_NN_KoELECTRA_Base.ipynb <==
# 준비
라이브러리, 파라미터 세팅

import pandas as pd
import numpy as np
import os
import re

DATA_IN_PATH = './data_in/'
DATA_OUT_PATH = './data_out/'

TEST_SIZE = 0.2
RANDOM_SEED = 42

# 데이터 불러오기

# 테스트(캐글) 데이터 로드
test_data = pd.read_csv(DATA_IN_PATH + 'ko_data.csv', encoding = 'cp949')
test_data.columns = ['id','document']  # 전처리 일괄 수행을 위해 컬럼명 변경(학습 데이터셋과 동일하게)
test_data.head()

# 로지스틱 회귀 예측 데이터 로드 (한국어 혐오 데이터 추가 학습)
predict_lr = pd.read_csv(DATA_OUT_PATH + 'NSMC_LogisticRegression_KHS.csv')
predict_lr.head()

# 버트 예측 데이터 로드
predict_bert = pd.read_csv(DATA_OUT_PATH + 'NSMC_BERT.csv')
predict_bert.head()

# GPT2 예측 데이터 로드
predict_gpt2 = pd.read_csv(DATA_OUT_PATH + 'NSMC_GPT2.csv')
predict_gpt2.head()

# CNN 예측 데이터 로드
predict_cnn = pd.read_csv(DATA_OUT_PATH + 'NSMC_CNN.csv')
predict_cnn.head()

# KoELECTRA(기본) 예측 데이터 로드
predict_koelectra = pd.read_csv(DATA_OUT_PATH + 'NSMC_KoELECTRA.csv')
predict_koelectra.head()

# KoELECTRA(기본+네이버쇼핑+스팀리뷰) 예측 데이터 로드
predict_koelectra_NaverShopping_Steam = pd.read_csv(DATA_OUT_PATH + 'NSMC_KoELECTRA_NaverShopping_Steam.csv')
predict_koelectra_NaverShopping_Steam.head()

# KoELECTRA(기본+감성사전) 예측 데이터 로드
predict_koelectra_SentiWordDict = pd.read_csv(DATA_OUT_PATH + 'NSMC_KoELECTRA_SentiWordDict.csv')
predict_koelectra_SentiWordDict.head()

# KoELECTRA(기본+감성사전+한국어혐오) 예측 데이터 로드
predict_koelectra_SentiWordDict_KHS = pd.read_csv(DATA_OUT_PATH + 'NSMC_KoELECTRA_SentiWordDict_KHS.csv')
predict_koelectra_SentiWordDict_KHS.head()

# KoELECTRA(기본+네이버쇼핑) 예측 데이터 로드
predict_koelectra_NaverShopping = pd.read_csv(DATA_OUT_PATH + 'NSMC_KoELECTRA_NaverShopping.csv')
predict_koelectra_NaverShopping.head()

# 기본 모델 (Baseline)
### 모델 앙상블(Majority Voting) 방식

# Id 컬럼 한 개를 살리기 위해 첫번째 데이터프레임을 제외한 나머지의 Id 컬럼만 삭제
#predict_koelectra_NaverShopping_Steam.drop('Id', axis=1, inplace=True)
predict_koelectra_NaverShopping.drop('Id', axis=1, inplace=True)
predict_koelectra_SentiWordDict.drop('Id', axis=1, inplace=True)
predict_koelectra_SentiWordDict_KHS.drop('Id', axis=1, inplace=True)

result = pd.concat([predict_koelectra,predict_koelectra_NaverShopping,predict_koelectra_SentiWordDict,predict_koelectra_SentiWordDict_KHS], axis=1)
#result = pd.concat([predict_lr,predict_bert,predict_gpt2,predict_koelectra,predict_cnn], axis=1)
result

result.columns = ['Id','KoELECTRA','KoELECTRA_NaverShopping','KoELECTRA_SentiWordDict','KoELECTRA_SentiWordDict_KHS']
#result.columns = ['Id','KoELECTRA','KoELECTRA_NaverShopping_Steam','KoELECTRA_SentiWordDict','KoELECTRA_SentiWordDict_KHS']
#result.columns = ['Id','Logistic_Regression','BERT','GPT2','KoELECTRA','CNN']
result.head()

result["Ensemble_Result"] = ((result["KoELECTRA"]+result["KoELECTRA_NaverShopping"]+result["KoELECTRA_SentiWordDict"]+result["KoELECTRA_SentiWordDict_KHS"])/3).astype('int')
#result["Ensemble_Result"] = ((result["KoELECTRA"]+result["KoELECTRA_NaverShopping_Steam"]+result["KoELECTRA_SentiWordDict"]+result["KoELECTRA_SentiWordDict_KHS"])/3).astype('int')
#result["Ensemble_Result"] = ((result["Logistic_Regression"]+result["BERT"]+result["GPT2"]+result["KoELECTRA"]+result["CNN"])/3).astype('int')
result.head(20)

output = result[['Id','Ensemble_Result']].copy()
output = output.rename({'Ensemble_Result':'Predicted'}, axis='columns')
output.head()

# 해당 경로가 없으면 생성
if not os.path.exists(DATA_OUT_PATH):
    os.makedirs(DATA_OUT_PATH)

# csv파일 생성
output.to_csv(DATA_OUT_PATH + "NSMC_Ensemble_KoELECTRA_Base_20201212.csv", index = False)  # 앙상블 결과, 캐글 제출 용도
#output.to_csv(DATA_OUT_PATH + "NSMC_Ensemble_KoELECTRA_Base.csv", index = False)  # 앙상블 결과, 캐글 제출 용도

### 캐글 제출 결과
**[2020.12.09]**<br>
Logistic Regression, BERT, GPT2, KoELECTRA 조합(4개) => 0.89647<br>
Logistic Regression, BERT, GPT2, KoELECTRA, CNN 조합(5개) => 0.89308<br>

**[2020.12.10]**<br>
KoELECTRA(에폭3)(0.89397),<br>
KoELECTRA_NaverShopping_Steam(에폭4)(0.89629),<br>
KoELECTRA_SentiWordDict(에폭3)(0.90112)(best),<br>
KoELECTRA_SentiWordDict_KHS(에폭3)(0.89379)<br>
조합(4개) => 0.90273<br>

KoELECTRA(에폭4)(0.89772)(best),<br>
KoELECTRA_NaverShopping_Steam(에폭3)(0.89236),<br>
KoELECTRA_SentiWordDict(에폭3)(0.90112)(best),<br>
KoELECTRA_SentiWordDict_KHS(에폭4)(0.89880)(best)<br>
조합(4개) => 0.90505

**KoELECTRA(에폭4)(best),<br>
KoELECTRA_NaverShopping_Steam(에폭4)(best),<br>
KoELECTRA_SentiWordDict(에폭3)(best),<br>
KoELECTRA_SentiWordDict_KHS(에폭4)(best)<br>
조합(4개) => 0.90577**

**[2020.12.12]**<br>
**KoELECTRA(에폭4)(best),<br>
KoELECTRA_NaverShopping(에폭4)(best),<br>
KoELECTRA_SentiWordDict(에폭3)(best),<br>
KoELECTRA_SentiWordDict_KHS(에폭4)(best)<br>
조합(4개) => 0.90631**
