==> 02_NSMC/12_NSMC_Ensemble_NN_KoELECTRA_SentiWordDict.ipynb <==
# 준비
라이브러리, 파라미터 세팅

import pandas as pd
import numpy as np
import os
import re

DATA_IN_PATH = './data_in/'
DATA_OUT_PATH = './data_out/'

TEST_SIZE = 0.2
RANDOM_SEED = 42

# 데이터 불러오기

# 테스트(캐글) 데이터 로드
test_data = pd.read_csv(DATA_IN_PATH + 'ko_data.csv', encoding = 'cp949')
test_data.columns = ['id','document']  # 전처리 일괄 수행을 위해 컬럼명 변경(학습 데이터셋과 동일하게)
test_data.head()

# KoELECTRA(기본) 예측 데이터 로드
predict_koelectra = pd.read_csv(DATA_OUT_PATH + 'NSMC_KoELECTRA.csv')
predict_koelectra.head()

# KoELECTRA(기본+감성사전) 예측 데이터 로드
predict_koelectra_SentiWordDict = pd.read_csv(DATA_OUT_PATH + 'NSMC_KoELECTRA_SentiWordDict.csv')
predict_koelectra_SentiWordDict.head()

# KoELECTRA(기본+감성사전+한국어혐오) 예측 데이터 로드
predict_koelectra_SentiWordDict_KHS = pd.read_csv(DATA_OUT_PATH + 'NSMC_KoELECTRA_SentiWordDict_KHS.csv')
predict_koelectra_SentiWordDict_KHS.head()

# KoELECTRA(기본+감성사전+네이버쇼핑) 예측 데이터 로드
predict_koelectra_SentiWordDict_NaverShopping = pd.read_csv(DATA_OUT_PATH + 'NSMC_KoELECTRA_SentiWordDict_NaverShopping.csv')
predict_koelectra_SentiWordDict_NaverShopping.head()

# KoELECTRA(기본+감성사전+스팀리뷰) 예측 데이터 로드
predict_koelectra_SentiWordDict_Steam = pd.read_csv(DATA_OUT_PATH + 'NSMC_KoELECTRA_SentiWordDict_Steam.csv')
predict_koelectra_SentiWordDict_Steam.head()

# 기본 모델 (Baseline)
### 모델 앙상블(Majority Voting) 방식

# 테스트 데이터의 id 부분을 리스트 처리
test_id = list(test_data['id'])
len(test_id)

# koelectra 데이터의 Predicted 부분을 리스트 처리
koelectra_predicted = list(predict_koelectra['Predicted'])
len(koelectra_predicted)

# koelectra_SentiWordDict 데이터의 Predicted 부분을 리스트 처리
koelectra_SentiWordDict_predicted = list(predict_koelectra_SentiWordDict['Predicted'])
len(koelectra_SentiWordDict_predicted)

# koelectra_SentiWordDict_KHS 데이터의 Predicted 부분을 리스트 처리
koelectra_SentiWordDict_KHS_predicted = list(predict_koelectra_SentiWordDict_KHS['Predicted'])
len(koelectra_SentiWordDict_KHS_predicted)

# koelectra_SentiWordDict_NaverShopping 데이터의 Predicted 부분을 리스트 처리
koelectra_SentiWordDict_NaverShopping_predicted = list(predict_koelectra_SentiWordDict_NaverShopping['Predicted'])
len(koelectra_SentiWordDict_NaverShopping_predicted)

# koelectra_SentiWordDict_Steam 데이터의 Predicted 부분을 리스트 처리
koelectra_SentiWordDict_Steam_predicted = list(predict_koelectra_SentiWordDict_Steam['Predicted'])
len(koelectra_SentiWordDict_Steam_predicted)

# 판다스 데이터프레임 통해 데이터 구성하여 output에 투입
result = pd.DataFrame( data={"Id": test_id, "KoELECTRA": koelectra_predicted \
                             ,"KoELECTRA_SWD": koelectra_SentiWordDict_predicted \
                             ,"KoELECTRA_SWD_KHS": koelectra_SentiWordDict_KHS_predicted \
                             ,"KoELECTRA_SWD_NaverShopping": koelectra_SentiWordDict_NaverShopping_predicted \
                             #,"KoELECTRA_SWD_Steam": koelectra_SentiWordDict_Steam_predicted \
                            } )
result

result["Ensemble_Result"] = ((result["KoELECTRA"]+result["KoELECTRA_SWD"]+result["KoELECTRA_SWD_KHS"]+result["KoELECTRA_SWD_NaverShopping"])/3).astype('int')
#result["Ensemble_Result"] = ((result["KoELECTRA"]+result["KoELECTRA_SWD"]+result["KoELECTRA_SWD_KHS"]+result["KoELECTRA_SWD_NaverShopping"]+result["KoELECTRA_SWD_Steam"])/4).astype('int')
#result["Ensemble_Result"] = ((result["Logistic_Regression"]+result["BERT"]+result["GPT2"]+result["KoELECTRA"]+result["CNN"])/3).astype('int')
result.head(20)

output = result[['Id','Ensemble_Result']].copy()
output = output.rename({'Ensemble_Result':'Predicted'}, axis='columns')
output.head()

# 해당 경로가 없으면 생성
if not os.path.exists(DATA_OUT_PATH):
    os.makedirs(DATA_OUT_PATH)

# csv파일 생성
output.to_csv(DATA_OUT_PATH + "NSMC_Ensemble_KoELECTRA_SentiWordDict.csv", index = False)  # 앙상블 결과, 캐글 제출 용도

### 캐글 제출 결과
**[2020.12.09]**<br>
Logistic Regression, BERT, GPT2, KoELECTRA 조합(4개) => 0.89647<br>
Logistic Regression, BERT, GPT2, KoELECTRA, CNN 조합(5개) => 0.89308<br>

**[2020.12.10]**<br>
KoELECTRA (에폭3)(0.89397),<br>
KoELECTRA_NaverShopping_Steam (에폭4)(0.89629),<br>
KoELECTRA_SWD (에폭3)(0.90112)(best),<br>
KoELECTRA_SWD_KHS (에폭3)(0.89379)<br>
조합(4개) => 0.90273<br>

KoELECTRA (에폭4)(0.89772)(best),<br>
KoELECTRA_NaverShopping_Steam (에폭3)(0.89236),<br>
KoELECTRA_SWD (에폭3)(0.90112)(best),<br>
KoELECTRA_SWD_KHS (에폭4)(0.89880)(best)<br>
조합(4개) => 0.90505

**KoELECTRA (에폭4)(best),<br>
KoELECTRA_NaverShopping_Steam (에폭4)(best),<br>
KoELECTRA_SWD (에폭3)(best),<br>
KoELECTRA_SWD_KHS (에폭4)(best)<br>
조합(4개) => 0.90577**

**[2020.12.11]**<br>
KoELECTRA (에폭4)(best),<br>
KoELECTRA_SWD (에폭3)(best),<br>
KoELECTRA_SWD_KHS (에폭4)(best)<br>
KoELECTRA_SWD_NaverShopping (에폭3)<br>
KoELECTRA_SWD_Steam (에폭3)<br>
조합(5개) => 0.90184

**KoELECTRA (에폭4)(best),<br>
KoELECTRA_SWD (에폭3)(best),<br>
KoELECTRA_SWD_KHS (에폭4)(best)<br>
KoELECTRA_SWD_NaverShopping (에폭3)<br>
조합(4개) => 0.90362**

감성사전 위주로 구성한 것보다 감성사전 포함한 것과 안한 것을<br>
적절히 섞는 경우 성능이 더 좋게 나옴(캐글 기준) 

**[2020.12.12]**<br>
**KoELECTRA(에폭4)(best),<br>
KoELECTRA_NaverShopping(에폭4)(best),<br>
KoELECTRA_SentiWordDict(에폭3)(best),<br>
KoELECTRA_SentiWordDict_KHS(에폭4)(best)<br>
조합(4개) => 0.90631 (Best!!)**
